reformat commands
