+++ src/abstract_commands/pools/delete_map.ts
