+++ src/abstract_commands/pools/delete_pool.ts
