+++ src/abstract_commands/pools/remove_map.ts
