+++ src/commands/pools/create_map.ts
@@ -1,12 +1,11 @@
 import {ApplicationCommandOptionTypes} from "discord.js/typings/enums";
 import {ICommand} from "wokcommands";
-
-import {removeMap} from "../../abstract_commands/maps/remove_map";
+import {createMap} from "../../abstract_commands/pools/create_map";
 import {ensure} from "../../utils/general";
 
 export default {
     category: "Admin",
-    description: "Remove a map from a pool",
+    description: "Create a map",
 
     slash: true,
     testOnly: true,
@@ -14,29 +13,31 @@ export default {
 
     options: [
         {
-            name: "pool_uuid",
-            description: "the uuid of the pool",
-            type: ApplicationCommandOptionTypes.INTEGER,
+            name: "name",
+            description: "The name of the map",
+            type: ApplicationCommandOptionTypes.STRING,
             required: true,
         },
         {
-            name: "map_uuid",
-            description: "the uuid of the map",
-            type: ApplicationCommandOptionTypes.INTEGER,
-            required: true,
+            name: "img_link",
+            description: "The link to an external image preview of the map",
+            type: ApplicationCommandOptionTypes.STRING,
+            required: false,
         },
     ],
 
     callback: async ({interaction}) => {
         const {options, channelId, guildId} = interaction;
+
         // ensure that the command is being run in a server
         if (!channelId || !guildId) {
             return "This command can only be run in a text channel in a server";
         }
-        // get the command parameters
-        const pool_uuid = ensure(options.getInteger("pool_uuid"));
-        const map_uuid = ensure(options.getInteger("map_uuid"));
 
-        return await removeMap(map_uuid, pool_uuid, guildId);
+        // get the command parameters
+        const name = ensure(options.getString("name"));
+        const imgLink = options.getString("img_link") ?? "";
+
+        return await createMap(name, imgLink, guildId);
     },
 } as ICommand;
