+++ src/commands/pools/create_pool.ts
@@ -1,14 +1,15 @@
 import {ApplicationCommandOptionTypes} from "discord.js/typings/enums";
 import {ICommand} from "wokcommands";
-import {createPool} from "../../abstract_commands/maps/create_pool";
+import {createPool} from "../../abstract_commands/pools/create_pool";
 import {ensure} from "../../utils/general";
 
 export default {
     category: "Admin",
-    description: "Create a Pool",
+    description: "Create a pool",
 
     slash: true,
     testOnly: true,
+    guildOnly: true,
 
     options: [
         {
