+++ src/commands/pools/list_pools.ts
@@ -10,6 +10,7 @@ export default {
 
     slash: true,
     testOnly: true,
+    guildOnly: true,
 
     options: [],
